+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Next
 
+## Breaking changes
+
+- Removed deprecated `OpenSnsTokenSwap`, `SetSnsTokenSwapOpenTimeWindow`, and `SetDefaultFollowees` from `ProposalActionRequest`.
+
 ## Features
 
 - Add support for `icrc21_canister_call_consent_message` to `@dfinity/ledger-icp` and `@dfinity/ledger-icrc`.
+- Add support for `"regtest"` in `BitcoinNetwork`.
 
 # 2024.09.02-0830Z